# Conflicts:
#	python/eulerlib.py
#	python/eulertest.py
#	python/p002.py
#	python/p003.py
#	python/p005.py
#	python/p006.py
#	python/p007.py
#	python/p009.py
#	python/p011.py
#	python/p012.py
#	python/p014.py
#	python/p016.py
#	python/p017.py
#	python/p019.py
#	python/p020.py
#	python/p022.py
#	python/p023.py
#	python/p024.py
#	python/p025.py
#	python/p030.py
#	python/p031.py
#	python/p032.py
#	python/p033.py
#	python/p034.py
#	python/p035.py
#	python/p037.py
#	python/p038.py
#	python/p041.py
#	python/p042.py
#	python/p043.py
#	python/p046.py
#	python/p047.py
#	python/p053.py
#	python/p054.py
#	python/p055.py
#	python/p059.py
#	python/p063.py
#	python/p069.py
#	python/p070.py
#	python/p072.py
#	python/p076.py
#	python/p077.py
#	python/p080.py
#	python/p091.py
#	python/p092.py
#	python/p116.py
#	python/p117.py
#	python/p124.py
#	python/p130.py
#	python/p132.py
#	python/p134.py
#	python/p135.py
#	python/p150.py
#	python/p162.py
#	python/p169.py
#	python/p182.py
#	python/p191.py
#	python/p216.py
#	python/p225.py
#	python/p267.py
#	python/p381.py
#	python/p387.py

+++ java/p022.java
@@ -1,6 +1,6 @@
 /* 
  * Solution to Project Euler problem 22
- * by Project Nayuki
+ * Copyright (c) Project Nayuki. All rights reserved.
  * 
  * https://www.nayuki.io/page/project-euler-solutions
  * https://github.com/nayuki/Project-Euler-solutions
@@ -16,20 +16,24 @@ public final class p022 implements EulerSolution {
 	}
 	
 	
+	/* 
+	 * We apply straightforward algorithms to sort the names, sum the letter values, and multiply by the position.
+	 */
+	
 	public String run() {
 		Arrays.sort(names);
-		int sum = 0;
+		long sum = 0;
 		for (int i = 0; i < names.length; i++) {
 			int value = 0;
 			for (int j = 0; j < names[i].length(); j++)
 				value += names[i].charAt(j) - 'A' + 1;
 			sum += value * (i + 1);
 		}
-		return Integer.toString(sum);
+		return Long.toString(sum);
 	}
 	
 	
-	private String[] names = {
+	private String[] names = {  // 10 strings per line, except final line. Mutable array
 		"MARY", "PATRICIA", "LINDA", "BARBARA", "ELIZABETH", "JENNIFER", "MARIA", "SUSAN", "MARGARET", "DOROTHY",
 		"LISA", "NANCY", "KAREN", "BETTY", "HELEN", "SANDRA", "DONNA", "CAROL", "RUTH", "SHARON",
 		"MICHELLE", "LAURA", "SARAH", "KIMBERLY", "DEBORAH", "JESSICA", "SHIRLEY", "CYNTHIA", "ANGELA", "MELISSA",
